requirements.txt
pymongo[srv]
flask
flask_restful
validators
lxml~=4.6.2
idna~=2.10
requests~=2.25.1
pip
python-dateutil~=2.8.1
six~=1.15.0
tornado~=6.1
pytz~=2020.5
Babel~=2.9.0
Jinja2~=3.0.1
Werkzeug~=2.0.1
itsdangerous~=2.0.1
click~=7.1.2
MarkupSafe~=2.0.1
packaging~=20.8
Pygments~=2.7.3
certifi~=2020.12.5
chardet~=3.0.4
urllib3~=1.26.2
aniso8601~=8.1.0
pyparsing~=2.4.7
setuptools~=47.1.0
netaddr~=0.8.0
redis~=3.5.3
rq~=1.7.0
rq_dashboard
python-dotenv
future
nose
publicsuffix>=1.1.0
py3dns>=3.1.0
pyspf==2.0.11
gunicorn
gevent
docopt
trustymail
python-nmap
colorama
py-emailprotections
dnslib
tldextract
argparse
aiodnsbrute
anubis-netsec
pydnsbl
python-whois=0.7.3
simplejson